Add core list repeater functionality (#149)
This PR adds the core functionality for a list repeater. Other items will add tests, allow `item` props for modules that are repeated, update module deletion behavior, and update the UI to match the mocks.

As part of this, refactored `ComponentTreePreview` so the `renderComponent` would be its own component, `ComponentPreview`. Some Studio Store actions were also refactored to consolidate code between pages and modules for component state updates.

J=SLAP-2639
TEST=manual

See that Studio is able to parse a page file with a `map` function, for example:
```tsx
{document.services.map((item, index) => (
  <Banner title={`${item}!`} key={index} />
))}
```
Also, see that a `map` can be created through the UI and written to file successfully.