Correct bugs in String Union type support.
Studio had two bugs when it came to String Union props. Firstly, it assumed
they were `PropValueKind.Expression`, which is incorrect. This assumption caused
Studio to break whenever a new value was selected from the Prop's drop-down. The
second bug was that the currently selected value did not appear first in the
drop-down.

TEST=auto,manual

Updated and re-ran all unit tests. Manually verified that I could change string
union values in the UI and that the currently selected value always appeared
first in the drop-down.