consolidate entity data into store (#156)
Removes useStreamDocument and similar code, and moves it into a single place in the store.
This avoids having the reload the same document over and over again.
Renames all usages of streamDocument to entityData for consistency, except for one helper in studio-plugin.
Removes setActivePageState because it was completely unused except in tests.

J=SLAP-2652
TEST=manual,auto

see that field picker works
see that switching pages changes the stream document
note that when you add a new page, it will not have an associated entity file since we don't know which files
it's able to accept. This has not changed.