Update temporal state and refactor prop input debouncing (#167)
This PR makes updates to the temporal state and refactors how debouncing is done for prop input values.

The snapshots of the store temporal state no longer include the `PreviousSaveSlice`, so undo/redo actions don't affect that information. When determining whether or not to take a snapshot, only a limited part of the state is compared for equality.

Debouncing of prop input values in the temporal state was moved into the component controlling the text inputs (`FieldPickerInput`) so that other temporal state updates weren't debounced. For example, previously, quickly clicking undo/redo would cause those updates to be throttled and mess up the temporal state. Similarly, store actions that updated multiple pieces of state (more than one `set` call), wouldn't always be captured properly for the temporal state.

Debouncing of the hex color prop values in the Studio store was also added to `PropInput`. This fixes the lag when clicking and dragging to change hex color in the color picker, because the actual value isn't updated in the store or component preview until the debounce time is over.

There are still a couple problems with undo/redo that would be fixed by current maintenance items we have:
- If a prop value is undefined, we set it to a default value as soon as the component is made the active component (and the `PropEditor` renders). This makes it impossible to undo past setting that component as active because undoing the default prop value causes it to be reapplied in an endless loop. Respecting undefined values would fix this ([item](https://yexttest.atlassian.net/browse/SLAP-2650)).
- Right now, the `pendingChanges` state is included in the control of the temporal store, which means an undo after a Save could create a state where Studio thinks there are pending page updates when there actually aren't. This would be fixed by moving that part of state into the `PreviousSaveSlice` ([item](https://yexttest.atlassian.net/browse/SLAP-2675)).

J=SLAP-2674, SLAP-2537
TEST=auto, manual

Click around in the test site and see that updating prop values and site settings and changing active components/pages/modules occurred as expected and could be undone/redone.