Conditionally show Field Picker and Repeater (#172)
Only show the Field Picker and Repeater when relevant entity data is present.

Currently, clicking on the Field Picker icon can lead to an empty dropdown being opened, which looks a little weird. Also, a Repeater editor appears for components even when there is no stream config for the page. Since arrays are only available through entity data for now, there's no way to actually set up the Repeater. In these cases, we shouldn't show the Field Picker icon or the Repeater editor.

Once we support other arrays (e.g. in site settings or props), we can remove the conditional check for displaying the Repeater editor.

J=SLAP-2679
TEST=manual, auto

See that the Field Picker icon and Repeater editor only appear on pages with entity data that match the expected field type (strings for props, arrays for Repeaters). Update the Playwright screenshots to show this behavior as well.