Add rearrange elements acceptance test (#175)
Add a Playwright test for rearranging the components in a component tree.

Playwright locators have a `dragTo` method, but it doesn't work in this case. We have to manually call the mouse up/down/move events because we need to be able to specify `steps` during `mouse.move`. This sends intermediate `mousemove` events during the move, which is needed for `@minoru/react-dnd-treeview` to register the dragging behavior properly.

J=SLAP-2482
TEST=auto