Ensure de-duplication when generating `fields` for the `TemplateConfig`.
This PR adds de-duplication logic to the `pagesJSFieldsMerger`. This ensures
that Studio always generates a unique set of `fields`, no field appears
twice. This issue came to light because someone was using two different elements
of an array field on their page as props. This resulted in the array field
appearing twice in `fields` whenever they saved.

TEST=auto

Added a test for the exact scenario the customer ran into. The test ensures we
are de-duping properly.