Create `ParsingError` type and add preliminary usages. (#181)
This PR introduces the `ParsingError` interface. This interface, which extends the built-in `Error`, is meant to convey any error that can occur in Studio's parsing code. The different types of parsing error are delineated by the `ParsingErrorType` enum.

I began adding `ParsingError`s to some parts of the code, replacing plain `throw new Error...` invocations. Later PRs will add more usages of the new error model. One thing to note is that I added `true-myth` to the repository. This library provides a `Result` interface which allows us to express that a method may yield a particular type of error. It's somewhat akin to being able to add `throws ...` in a Java function signature.

J=SLAP-2684
TEST=auto

Updated and re-ran unit tests.