Support parsing and writing simple getPath functions (#183)
Add support for parsing simple `getPath` functions and writing them to file. Currently, only `getPath` functions that return a single, top-level, string literal are parsed. This is because static pages don't have a stream config, so they don't have access to `document` fields, and in SLAP-2701, users will only be able to specify simple return values rather than code when adding a new page.

J=SLAP-2700
TEST=auto, manual

Added a static page to the test-site with a simple `getPath` function. See that on Save, it is written back to file with no changes.