Add Syntax Validation to PageFile.getPageState (#185)
This PR adds syntax validation to getPageState via the babel parser.
I opted for the babel parser instead of calling the typescript compiler we don't need to do type checking on
the file and all of its dependencies, and the babel parser should also be much faster because of that.
This is just to prevent the majority of syntax errors that would not be caught during parsing, but would cause
the page to display an error on load.

Also moves createStudioPlugin from the index file into its own entrypoint so that it doesn't get imported into the frontend.
This lets us remove all the annoying dynamic imports we were doing for packages that would error out if the frontend tried
to import them (babel was one of these). 
This also removes all the assorted different network requests for files like HmrManager or packages like simple-git, which were also being imported by the frontend due to being dependencies of createStudioPlugin.

Also remove NpmLookup from the index entrypoint, removing an 8.5Mb network request for the entire typescript package. In the future we'll want to bundle together the index entrypoint, so that it the 10Mb ts-morph import
can be tree shaken down to the only bit we need from it - `SyntaxKind`.

J=SLAP-2706
TEST=manual,auto

see that I will get a console error if my page has invalid syntax