Support static page creation (#186)
Add support for creating static pages in PagesJS repos. When `isPagesJSRepo` is true in the Studio config, the Add Page button asks for both a name and URL slug for the page before the page can be created.

J=SLAP-2701
TEST=auto, manual

Add a page in the test-site and see that a URL slug value is required. On Save, see that the new page file is written with a `getPath` function with the correct return value.