Improve error handling when Studio Config is invalid. (#191)
This PR improves the error handling that occurs when a Studio Config import fails
or the Config is not a `default` export. In both cases, we still exit the application
in the terminal. But, there are now pretty-printed messages indicating why Studio had
to shut down.

J=SLAP-2705
TEST=auto, manual

Added and updated unit tests. Tried a Studio starter out with these changes. Saw the
appropriate error messages when Studio crashed due to a bad config. Also saw the Studio
app start up correctly for a valid Config.