Add basic page settings editor (#192)
Add a button next to each page name for PagesJS repos. Clicking on the button opens an editor modal for the template meta info, which is just `getPath` value for now. Right now, only string literal `getPath` values are supported, but the next PR(s) will add support for template literals and references to entity data.

J=SLAP-2712
TEST=auto, manual

In the test-site, saw the settings button appear next to each page name. For `UniversalPage`, where the `getPath` value is more complex than just a string literal, the button is disabled and shows a tooltip on hover. When `isPagesJSRepo = false` in the config, the settings buttons are no longer rendered.