ErrorComponentPreview UI (#194)
This PR adds UI for previewing ErrorComponentStates.
On hover, components are given a red outline, a tooltip with the error, and a small box shadow.
We attempt to import components even if they could not be parsed, since rendering them does not necessarily require type information. If there is an error during the render, that error is displayed in place of the component.

The active component highlighting displays a similar red outline instead of a blue outline for ErrorComponentStates.

Fixes a bug that was preventing acceptance test snapshots from being updated correctly.

--- 
### Updates our playwright.yml workflow to only run tests on push, and not pull_request.
While trying to iron out the playwright snapshot infrastructure, I noticed that running tests on push vs pull_request will actually result in UI, thanks to how the deploy button reads the git state to detect whether or not a commit can be pushed.

When a workflow runs on pull_request, github creates a merge commit between the head ref (here dev/errorcomponentstate-ui) and the base ref (here the `main` branch).
On this specific merge commit, there is no upstream branch, so the deploy button will always be grayed out.

When a workflow runs on push, it runs directly on the pushed commit for the head branch, i.e. dev/errorcomponentstate-ui.
Because this branch DOES have an upstream, when changes are saved, the deploy button will become active.

It's also worth noting that there can also be a mismatch between running snapshots locally and running them on a `push` workflow. If the local branch has changes the upstream does not, the deploy button will be active on load. However when run on a `push` workflow, the local branch will never have changes the upstream doesn't, so the deploy button will always be disabled on load.
I created SLAP-2719 to address this issue since it's pretty annoying and makes the snapshots inconsistent, with Playwright snapshots already being somewhat inconsistent even without this issue.

J=SLAP-2709
TEST=manual