Support editing getPath expressions in UI (#198)
Update the template settings editor so that `getPath` values are treated as expressions, similar to string prop inputs. As for props, square or curly brackets can be used to reference entity data. Curly braces in the input are immediately converted to square brackets. Unlike prop inputs, the `getPath` input does not include a field picker.

Also, the logic for computing the fields in the stream config is updated to look for stream document paths in the `getPath` value as well as the component tree.

J=SLAP-2714
TEST=auto, manual

Successfully update the `getPath` value for `UniversalPage` and `index` in the test-site.