Support parsing and writing stream scope (#204)
This PR adds support for the parsing and writing-to-file of an entity template's stream scope.

Since entity templates must have a stream scope defined, we can use this field to differentiate between static and entity pages. Now, a template config is only added/updated for a page file if it's an entity page, i.e. it has a stream scope defined, instead of based on whether a `document.blah` expression is present on the page.

I also updated the page settings editor to only allow expressions for an entity page's `getPath` value, since static pages don't have a `document` or other expression sources that they can reference. The value for a static page is treated as a string literal. If an expression is returned by the `getPath` value of a static page, it is no longer editable in the UI.

J=SLAP-2732
TEST=auto, manual

Check that editing the `getPath` value of a static page writes a string literal to file, while an entity pages writes an expression.