Add modal for stream scope (#206)
Add a modal in the page creation flow for specifying the scope of the stream for entity pages. None of the fields for the stream scope are required, so all fields in the modal are optional. String input values are considered to be comma-separated lists.

This PR also includes a small fix so that the `document` param isn't added for a page component if it doesn't use any stream expressions.

J=SLAP-2728
TEST=auto, manual

In the test-site, saw that the stream scope can only be specified in a PagesJS repo when the entity page type is selected. Tested specifying stream scope with zero, one, and multiple filter values.