Allow getPath expression during entity page creation (#207)
In the entity page creation flow, update the modal for specifying the page name and URL so that the URL field is auto-populated with `[[slug]]` and values entered in that field are treated as expressions.

J=SLAP-2729
TEST=auto, manual

Add static and entity pages. See that the values are treated as literals for static pages and expressions for entity pages.