Add UI for array prop expressions (#216)
Add the UI for previewing and editing expressions for array props. The prop input UI is similar to that of string props, but selecting an item via the field picker completely replaces the expression value for the prop, rather than adding it onto the end as it is for strings.

J=SLAP-2744
TEST=manual, auto

Successfully edit the `words` array prop in the Banner component in the test-site.