Support Deploy within CBD (#220)
This PR adds necessary support for the DeployButton within CBD.
After a recent CBD change, there is a new environment variable, YEXT_CBD_BRANCH, which contains the git ref
for the current branch being edited in CBD. Studio will now switch to that branch (instead of the default `livepreview` branch that CBD automatically creates) so that git pushes can be made.

CBD is missing user.email and user.name git config which is necessary for pushing to git, so those are now provided to simple-git via the config param in the SimpleGitFactory. Providing the config this way will prefix all commands made by simple-git with inline config. Doing this via inline config is good because it won't pollute the global git config, which can be annoying for local development of studio or developers who use studio locally (i.e. not via a container).

The SaveButton is also hidden when studio runs via CBD. The technical reason for this is that, without websocket support in CBD, the gitData virtual module does not accurately reflect the state of the repo, which causes behavior where clicking on the save button will disable the deploy button due to the gitData not updating after files are saved to the repo.
The product-y reason for this is that the save button is much less useful in CBD, since saved files only persist while your CBD session persists, and could also be confusing for new users to studio.
I decided to just do this change for now and see what Andrew says when he gets back.

J=SLAP-2748
TEST=manual

Locally I can deploy and it won't mutate my git config
I can also hit save and THEN still hit deploy, the deploy won't be disabled
the save button is hidden in CBD's live preview
I can deploy multiple times in a row in CBD
I can open any arbitrary branch in CBD
.initializationdata is not committed in CBD
no error is thrown if .initializationdata does not exist in the repo