Address UI Feedback for Complex, Editable Props. (#225)
Andrew noted that the left-border for an object-type Prop extends the entire length of its `NestedPropEditors`. This is not desired, we want UI that is more sleek. Now, the left-border extends only a small bit below this Prop's name.
For the last element within an object-type Prop, provided it's an atomic type, the left border makes a perfect right-angle with the horizontal line extending from the prop name.

Please see the re-generated `renders-nested-props-1` screenshot for a before and after comparison.

J=SLAP-2756
TEST=manual, auto