Windows Support: Remove colon from virtual module paths (#230)
This PR updates the virtual module import specifier from `virtual:yext_studio` to `virtual_yext_studio`. 
On windows, colons are not allowed as parts of filenames, and trying to clone a repo that has a filename with a colon on it will fail in fantastic manner. Since we had a `__mocks__/virtual:yext_studio` file, this would prevent the repo from being cloned on windows. 

J=SLAP-2776
TEST=manual

I can clone the repo on windows