Windows Support: can run build in windows (#231)
`rm` is not a valid command in windows terminal/powershell, so it's been replaced with rimraf.
`chmod` is also not a valid command in windows. There's also no equivalent due to windows using a different permissions scheme. Instead, I added `run-script-os` to the repo, and we simply don't run chmod on windows.

`chmod` is necessary in the first place because when the typescript compiler spits out new files, it will never mark them as executable because it's "beyond the scope of typescript" https://github.com/microsoft/TypeScript/issues/16257
It will however preserve the permissions when running in watch mode, which is why we don't typically need to run chmod while doing normal development.

J=SLAP-2775
TEST=manual,auto

the build test job is now run on windows
I can do local development of studio in windows