Support previewing and editing existing array literals (#232)
Add support for previewing and editing existing array literal values. Now, if an array prop has a literal value, an editor for each item in the array is displayed below the expression input (see screenshot). The expression input is disabled if the value is a literal and there are items in the array. Support for adding new items to the array and removing existing items will be added in future PRs.

Also, fixed some bugs:
- Add back tooltips for prop docs and add testing to prevent future regressions
- When updating the stream fields in the template config, look through object prop fields for stream expressions
  - Right now, the logic in `ComponentTreeHelpers.usesExpressionSource` and `TemplateConfigWriter.getUsedStreamDocumentPaths` are separate but similar. Since this PR was already getting fairly big, I made a separate item (SLAP-2780) to refactor them so the code can be reused

J=SLAP-2773, 2767
TEST=auto, manual

Successfully edit expression and literal array prop values.

<img width="313" alt="Screen Shot 2023-06-14 at 7 19 57 PM" src="https://github.com/yext/studio-prototype/assets/88398086/c3459c51-002c-49f8-8969-b6e961c716a3">