Validation for new page name and overflow (#222)
This PR adds errors for when the user inputs a page name that is not a valid filename (has certain special characters, is too long). If the page name is a valid filename but not a valid React component name (doesn't start with a capital letter, has special characters and whitespace), it edits the name to be a valid React component in reactComponentNameSanitizer() but doesn't change the name that is displayed on Studio. Also allows page names to overflow in the active page panel when the string is too long.

J=SLAP-258
TEST=auto, manual

Tested different page names ("test...", "test/test", etc) to make sure they threw errors.
Tested different page names ("pa-ge", "pa<>---ge", "9999", etc) to make sure they resulted in an edited component name.