Windows Support: NpmLookup Changes (#233)
This PR updates NpmLookup to work with windows. This PR doesn't really have windows specific changes but addresses some issues with NpmLookup that would only surface on Windows (but still should be addressed on mac/linux).

NpmLookup works fine when trying to resolve a node module, since ModuleResolutionHost will end up being passed a relative path. However, it did not work when trying to resolve imports to local files on windows.
This PR streamlines some of the logic, removes some idiosyncrasies, and by doing so fixes the windows related issues.

J=SLAP-2783
TEST=manual,auto

the exportedTypes import in Banner.tsx loads correctly, before we would get a (gracefully handled) ParsingError for it

### NpmLookup details
The way NpmLookup works is that, given some importSpecifier like `@yext/search-ui-react`, in order to resolve the source file path for this import (so that we can parse it using ts-morph), we would pass this importSpecifier into `typescript.resolveModuleName`. `typescript.resolveModuleName` has some logic to know that we're looking for a node module (likely by seeing that the import is not a relative path), so it would then look for file patterns like `node_modules/@yext/search-ui-react/index.ts`,
`node_modules/@yext/search-ui-react/index.tsx`,
`node_modules/@yext/search-ui-react/index.d.ts`,
`node_modules/@yext/search-ui-react/package.json`, 
`node_modules/@yext/search-ui-react/index.js`, etc. (if curious it's easy to see the list by inserting a console.log into our customModuleResolutionHost)
until it finds a match. The way it's able to tell whether a match is found is via the `ModuleResolutionHost` we pass into it, which we use to define `fileExists` and `readFile` methods which tell `typescript.resolveModuleName` whether or not a file like one of the above listed exist.
Since the filepaths above are all relative paths, we do a `path.join` with a rootpath that beings at`process.cwd()` and the filepath, and then check/read the fully joined path.
In order to support monorepos, if no match is found, we set the rootpath to its own parent, and repeat the process.

This works fine for node modules since our resolution host gets passed a relative path at the end. But this didn't work (on windows) when trying to resolve an import to a local file, like the `import { NestedProp } from "../types/exportedTypes";` we have in our test-site.
When resolving modules, before this change NpmLookup would set the initial root to process.cwd(), and then continually move up one level (i.e. add `..` to its path). The cwd is set to the directory the studio process is started in, for example `studio-prototype/apps/test-site` when we run our test-site. it's not immediately obvious, but this causes issues when trying to import relative paths. Say you have a component located in `test-site/src/components/Banner.tsx` that does a relative import for `../types/exportedTypes` (yes this is literally an example using the current test-site). If you start from a cwd of `test-site` and look for `../types/exportedTypes`, you won't find it, because the root is already too deep. Instead, we need the root to start out at the path for `Banner.tsx`. On top of that, it should not recursively look for `../types/exportedTypes`, since in certain unlucky scenarios that could actually fetch the wrong file, and unlike when looking for a specific node module in a nested monorepo, there's no use case here.

The way we got around this previously was by turning the relative import path into an absolute one, before passing it into NpmLookup This worked in a rather janky way - it would still do a path.join for the current root directory and this absolute path, but eventually would be able to find a match once the root directory hit the value `/`, and it would do a `path.join('/', '/Users/blah/studio-prototype/apps/test-site/src/types/exportedTypes.ts`, which would end up with the correct path. This was not ideal but worked for most circumstances barring extremely weird edge cases.
However this doesn't work on windows since the true root path of the computer is something like `c://`, so at the very end it would do a path.join and end up with something like `c://c://blah` which obviously isn't valid.

Instead, it makes more sense to set the initial root path to something like `Banner.tsx`. After doing this, we don't need to turn the relative import path into an absolute one, and everything falls into place.
This also lets us get rid of some of the esoteric logic in NpmLookup .
We also shouldn't recurse if importing from a relative path.