Fix syncing of error page states (#249)
Update the HMR of the Studio store so it is synced with the latest version of the error page states in the file system. 

Also, add an acceptance test to check for this by ensuring that a page in an error state is removed from the list of error pages and is preview-able as a regular page once the errors in the file are fixed. All the snapshots are updated to include a new `ErrorPage`.

J=SLAP-2766
TEST=auto, manual

See that returning JSX from the `ErrorPage` in the test-site and adding a `getPath` function correctly updates the Studio UI via HMR.