Correct width of Drag Placeholder. (#251)
In the `ComponentTree`, we had incorrectly styled the Drag Placeholder. As you moved through the layers of the tree, we correctly increase the `left` CSS property. But, we were not decreasing the width of the Placeholder accordingly. This meant the red line associated with the Placeholder would stick out into the Page Preview for sufficiently nested Components.

This PR ensures that whenever we indent the Placeholder to the left, we reduce its width by the same amount.

J=SLAP-2800
TEST=manual