Allow Editing of Page's Stream Config (#236)
This pull request adds the ability to edit the stream scope in the page settings button, as well as instructions that such edits may cause entity data to become undefined. In order to accomplish this, it creates a new PageSliceAction to update the streamScope. 

It also changes the behavior when the getPath function returns undefined. Previously, when the path is undefined, the URL settings were not able to be edited via the UI. Now, a message says that the URL slug is defined by the developer and the URL settings are editable. 

J=SLAP-2723
TEST=auto,manual

Tested that three new form fields were created in the modal of the page settings button for entity pages and not static pages.
Tested that saving the modal causes a temporary save in Studio but does not change the local copy.
Tested that saving in Studio causes the temporary stream scope changes to be saved in local.
Tested that the above also works with empty fields.