Respect undefined prop values (#256)
Update prop previewing and editing to respect undefined values for optional props. `undefined` is a valid value for optional props and should not be modified by Studio unless the user explicitly makes that change. If a required prop has a value of `undefined` passed to it, the page will have a TS error. For now the prop editors look the same for an undefined value and the prop type's default value. The UI will be updated in SLAP-2799.

Also fixed a couple bugs:
- When new components are added to the page via the UI, we wouldn't add any values for the props, even if the props were required. So on Save, the page file would have a TS error. I updated the component creation action to populate required props with a default.
- Previously, we would use empty string as the default value for string unions, which is (usually) incorrect and would result in a TS error on Save. I updated the default value for string unions so it chooses from one of the union values.

This also fixes some of our undo/redo functionality which could enter an infinite loop when editing a prop with an undefined value. We would automatically set the value to the default, so when a user tried to Undo, the value would switch back to undefined, but then we'd automatically set the value back to the default. So they'd never be able to Undo past that point and it could lead to the history being in a weird state. This is no longer a problem now that we leave prop values as is instead of forcing them to default.

J=SLAP-2650
TEST=auto, manual

See that undefined prop values in the test-site that previously showed a default value in the preview will now show "undefined".