Add Acceptance Test for Static Page Creation (#257)
This pull request modifies the existing add-and-remove-page acceptance test to work with a PagesJS repo and to test removing and adding a static page. It also modifies the other acceptance tests to work with a PagesJS repo (mostly by adding GetPath functions).

J=SLAP-2736
TEST=manual

Watched videos to confirm that tests are working correctly.