Add Acceptance Test for Entity Page Creation (#262)
This pull request adds an acceptance test for adding an entity page. It also renames the test "can remove a page and then add it back" to "can remove a static page and then add it back" and regenerates its screenshots.

J=SLAP-2772
TEST=manual

Checked screenshots to confirm that test is working correctly.