Use Bespoke Studio Instances for Acceptance Tests (#263)
This PR updates the acceptance test infra to run in bespoke temporary directories, instead of juggling between different git branches. This lets tests be more isolated, and solves most issues of tests bleeding into each other, which previously was an issue with the `add and remove page` test. It also avoids annoying funky states with the git branches not cleaning up properly, especially when manually cancelled via something like `ctrl + c` which basically guaranteed that something weird would happen.
It also lets us run tests in parallel, which I thought would improve test speed. This is true when the tests are run locally, since I believe we all have machines with at least 12 cores (that's how many I have). By playwright's recommendation this lets me use 6 workers.

However, on it's own this change actually makes the tests run slower in CI, taking about 7 minutes, as opposed to about 2-3 minutes. This is only including the time to actually run the tests, and not any setup like installing browsers or npm ci. 
This is because the github test runners only have two logical cores, which is a scenario where running tests in parallel doesn't make sense. By default, when `fullyParallel` is turned on in the playwright config, playwright will assign one worker per 2 logical cores, leading to only 1 worker being run in CI for us. I tried manually specifying 2 workers, and it actually caused the tests to run slower (14+ minutes).
Without parallelization, each test now takes longer due to needing to start its own studio instance.

To remedy this for CI I added some sharding that lets us break up the tests into multiple github test runners.
https://playwright.dev/docs/test-parallel#shard-tests-between-multiple-machines.

I want to add github workflow caching for things like npm dependencies, installing playwright browsers, and hopefully even the studio build, which should speed things up across the board, and also cut down on github runner costs (which I have no context on how much we're paying or even what plan we're on).

---

The deploy test is temporary commented out. A followup PR will add it back in, with the plan being to update the related infra to push to a bespoke test repo, instead of the actual studio-prototype repo. This will make it harder for us to shoot ourselves in the foot with infinite github workflows (since this bespoke repo won't have any) and also remove the last bit of git branch juggling from the infra.

---

The following playwright source files were used for reference for things like waiting for the studio port to be ready.

https://github.com/microsoft/playwright/blob/608e336dba8f5ba4581069ed208103960712b783/packages/playwright-core/src/utils/processLauncher.ts

https://github.com/microsoft/playwright/blob/608e336dba8f5ba4581069ed208103960712b783/packages/playwright-test/src/plugins/webServerPlugin.ts

For waiting on the port to be ready, I tried the wait-on and wait-for-localhost packages. However neither of these seemed to work with the vite server.

J=SLAP-2805
TEST=auto

tests pass locally with full parallelism