Bespoke Git Instance for Deploy Test (#264)
This PR updates the createRemote flag to create a bespoke git repo for the deploy test.
This avoids issues with the test polluting the actual repo, and weird states when the test is cancelled prematurely
before cleanup is allowed to run.
It also avoids some of the cases that can cause infinite looping github workflows, since the commit that the deploy test
pushes to remote will no longer include any of the .github files from the actual repo.
This also avoids issues with the state of the actual repo bleeding into the test, which would cause inconsistencies with
the deploy button's state, since the deploy button depends on state read from the current repo.
Since we no longer have this inconsistency, we can now run the acceptance tests on the pull_request event instead of the push event. This is preferable for, again, preventing issues with infinite looping github workflows and overall results in us running acceptance tests in unnecessary situations, like somebody pushing incremental commits to a dev branch, but before that dev branch is ready for code review.

J=SLAP-2805
TEST=auto