Pass Preview into IFrame via React Portal (#278)
This PR updates the preview to be passed into an iframe via react portal.
This makes things like fixed styling work as expected within the iframe.

I also added an e.preventDefault() which will prevent previewed pinks from being navigable.
This is preferable over turning off pointer-events completely so that styling still appears as expected.

react-tooltip does not work out of the box inside an iframe, so we have to manually tell it the correct
coordinates to render at. It's also impossible for an iframe to render overflow content, so to render
the tooltip over the preview like we do in one of our screenshots, I shifted the contents of the iframe down
by 50px to make extra space for the tooltip.
https://github.com/yext/studio-prototype/blob/e5318a3f941543db21d101bb474e29fe675a166f/e2e-tests/__screenshots__/error-component-preview.spec.ts/components-with-parsing-errors-2.png
In order to do this I had to update react-tooltip to get access to new props.
However, in 5.7.5 react-tooltip made it so that when the tooltip is not open, it is no longer rendered to the DOM,
as opposed to having visibility hidden. This broke our unit tests so I had to update those.
https://github.com/ReactTooltip/react-tooltip/pull/932

J=SLAP-2821
TEST=manual

I can click on a component within the preview (which is now in an iframe) and the active component will be updated
component highlighting still works as expected.